Updated samples and docstrings
